+++ tests/test_engine.py
@@ -12,12 +12,14 @@ if os.path.exists('/var/www/trade-exec-engine/resources/test'):
 else:
     REPO_DIR = '/Users/pralhad/Documents/99-src/98-trading/trade-exec-engine'
 
-os.environ["ACCOUNT"] = "Trader-V2-Pralhad"
+ACCT = "Trader-V2-Pralhad"
+os.environ["ACCOUNT"] = ACCT
 os.environ["GENERATED_PATH"] = os.path.join(REPO_DIR, "generated")
 os.environ["LOG_PATH"] = os.path.join(REPO_DIR, "logs")
 os.environ["RESOURCE_PATH"] = os.path.join(REPO_DIR, "resources/config")
 
 from exec.service import engine
+from exec.utils.ParamBuilder import load_params
 
 sm = engine
 
@@ -51,28 +53,29 @@ class TestEngine(unittest.TestCase):
         mock_api.return_value = mock_response
         self.sm = sm
         self.sm.api.api_login()
+        self.sm.params = load_params(api=mock_api, acct=ACCT)
 
-    @patch.dict('exec.service.engine.cfg', {"generated": os.path.join(TEST_RESOURCE_DIR, 'load_params')})
-    @patch('exec.service.engine.api.api_get_order_book')
-    def test_load_params(self, mock_api):
-        """
-        Scenarios:
-        1. Only Entries file and empty OB
-        2. Entries plus partial OB - only partially Open
-        :param mock_api:
-        :return:
-        """
+    # @patch.dict('exec.service.engine.cfg', {"generated": os.path.join(TEST_RESOURCE_DIR, 'load_params')})
+    # @patch('exec.service.engine.api.api_get_order_book')
+    # def test_load_params(self, mock_api):
+    #     """
+    #     Scenarios:
+    #     1. Only Entries file and empty OB
+    #     2. Entries plus partial OB - only partially Open
+    #     :param mock_api:
+    #     :return:
+    #     """
+    #
+    #     mock_response = read_file("load_params/open-order-book.json")
+    #     mock_api.return_value = mock_response
+    #
+    #     sm.load_params()
+    #     result = read_file_df("load_params/expected-params.json")
+    #     result['target_pct'] = np.NaN
+    #
+    #     pd.testing.assert_frame_equal(sm.params, result)
 
-        mock_response = read_file("load_params/open-order-book.json")
-        mock_api.return_value = mock_response
-
-        sm.load_params()
-        result = read_file_df("load_params/expected-params.json")
-        result['target_pct'] = np.NaN
-
-        pd.testing.assert_frame_equal(sm.params, result)
-
-    @patch.dict('exec.service.engine.cfg', {"generated": os.path.join(TEST_RESOURCE_DIR, 'order_update')})
+    @patch.dict('exec.utils.ParamBuilder.cfg', {"generated": os.path.join(TEST_RESOURCE_DIR, 'order_update')})
     @patch('exec.service.engine.api.api.single_order_history')
     @patch('exec.service.engine.api.api_get_order_book')
     def test_event_handler_order_update(self, mock_api, order_hist_api):
@@ -96,9 +99,9 @@ class TestEngine(unittest.TestCase):
         mock_response.return_value = None
         mock_api.return_value = mock_response.return_value
         order_hist_api.return_value = None
-        sm.load_params()
 
-        # 1. New BO Creation 10 Order Updatas
+        # 1. New BO Creation 10 Order Updates
+        self.sm.params = load_params(api=mock_api, acct=ACCT)
         recs = read_file("order_update/1-bo-entry-order-update.json")
         for message in recs:
             sm.event_handler_order_update(curr_order=message)
@@ -109,7 +112,7 @@ class TestEngine(unittest.TestCase):
         pd.testing.assert_frame_equal(output, expected_params)
 
         # 2. SL Update - Successful
-        sm.load_params()
+        self.sm.params = load_params(api=mock_api, acct=ACCT)
         # Create Successful BO
         recs = read_file("order_update/1-bo-entry-order-update.json")
         for message in recs:
@@ -124,7 +127,7 @@ class TestEngine(unittest.TestCase):
         pd.testing.assert_frame_equal(output, expected_params)
 
         # 3. SL Update - Failed
-        sm.load_params()
+        self.sm.params = load_params(api=mock_api, acct=ACCT)
         # Create Successful BO
         recs = read_file("order_update/1-bo-entry-order-update.json")
         for message in recs:
@@ -146,7 +149,7 @@ class TestEngine(unittest.TestCase):
         pd.testing.assert_frame_equal(output, expected_params)
 
         # 4. SL Hit
-        sm.load_params()
+        self.sm.params = load_params(api=mock_api, acct=ACCT)
         # Create Successful BO
         recs = read_file("order_update/1-bo-entry-order-update.json")
         for message in recs:
